add skipping for building /copying /icon setting in cocos3

+++ vars/copyCocosBuildToJenkins.groovy
@@ -6,7 +6,11 @@ def call(Map args = [:]) {
     if (!unityProjectPath || !cocosProjectPath || !cocosVersion) {
         error "❌ 'unityProjectPath', 'cocosProjectPath', and 'cocosVersion' are required"
     }
-
+    // 💡 Skip condition
+    if (params.ENVIRONMENT == 'Production' && args.cocosVersion == 'cocos3') {
+        echo '⏭️ Skipping copying for cocos3 in Production mode.'
+        return
+    }
     def productName = sh(
         script: "grep 'productName:' '${unityProjectPath}/ProjectSettings/ProjectSettings.asset' | sed 's/^[^:]*: *//'",
         returnStdout: true
@@ -25,24 +29,32 @@ def call(Map args = [:]) {
             rm -rf '${sourceBuildFolder}'
         """
 
-        echo "✅ Cocos 2 build copied and original build folder deleted."
+        echo '✅ Cocos 2 build copied and original build folder deleted.'
         return targetBaseFolder
-
     } else if (cocosVersion == 'cocos3') {
         def sourceFullProject = cocosProjectPath
         echo "📂 Copying full Cocos 3 project from ${sourceFullProject} to ${targetBaseFolder}"
 
-        sh """
+        if (params.ENVIRONMENT == 'Testing') {
+            sh """
             mkdir -p '${targetBaseFolder}'
             cp -R '${sourceFullProject}/.' '${targetBaseFolder}/'
             cd '${targetBaseFolder}'
             find . -mindepth 1 -maxdepth 1 ! -name 'build' ! -name 'native' -exec rm -rf {} +
         """
+            echo '✅ Cocos 3 project copied and cleaned (kept only build and native folders).'
+        } 
+        else {
+            sh """
+            mkdir -p '${targetBaseFolder}'
+            cp -R '${sourceFullProject}/.' '${targetBaseFolder}/'
+        """
+            echo '✅ Cocos 3 project copied (all files and folders kept).'
+        }
 
-        echo "✅ Cocos 3 project copied and cleaned (kept only build and native folders)."
         return targetBaseFolder
-
-    } else {
-        error "❌ Unsupported cocosVersion: ${cocosVersion}"
     }
+ else {
+        error "❌ Unsupported cocosVersion: ${cocosVersion}"
+ }
 }
